ref(endpoints): Split query endpoints
