+++ server/src/endpoints/project_configs.rs
@@ -1,28 +1,11 @@
 use std::rc::Rc;
 
-use actix_web::http::Method;
-use actix_web::Error;
-use actix_web::Json;
-
+use actix_web::{http::Method, Error, Json};
 use futures::{future, Future};
 
-use actors::keys::{GetPublicKeys, GetPublicKeysResult};
 use actors::project::{GetProject, GetProjectState, GetProjectStates, GetProjectStatesResponse};
-use extractors::CurrentServiceState;
-use extractors::SignedJson;
+use extractors::{CurrentServiceState, SignedJson};
 use service::ServiceApp;
-
-#[cfg_attr(feat = "cargo-clippy", allow(needless_pass_by_value))]
-fn get_public_keys(
-    (state, body): (CurrentServiceState, SignedJson<GetPublicKeys>),
-) -> Box<Future<Item = Json<GetPublicKeysResult>, Error = Error>> {
-    let res = state.key_manager().send(body.into_inner());
-
-    Box::new(
-        res.map_err(Error::from)
-            .and_then(|x| x.map_err(Error::from).map(|x| Json(x))),
-    )
-}
 
 #[cfg_attr(feature = "cargo-clippy", allow(needless_pass_by_value))]
 fn get_project_configs(
@@ -63,9 +46,7 @@ fn get_project_configs(
 }
 
 pub fn configure_app(app: ServiceApp) -> ServiceApp {
-    app.resource("/api/0/relays/publickeys/", |r| {
-        r.method(Method::POST).with(get_public_keys);
-    }).resource("/api/0/relays/projectconfigs/", |r| {
+    app.resource("/api/0/relays/projectconfigs/", |r| {
         r.method(Method::POST).with(get_project_configs);
     })
 }
