+++ server/src/service.rs
@@ -77,10 +77,11 @@ fn make_app(state: ServiceState) -> ServiceApp {
         .middleware(AddCommonHeaders)
         .middleware(ErrorHandlers);
 
-    app = endpoints::healthcheck::configure_app(app);
-    app = endpoints::store::configure_app(app);
     app = endpoints::forward::configure_app(app);
-    app = endpoints::queries::configure_app(app);
+    app = endpoints::healthcheck::configure_app(app);
+    app = endpoints::project_configs::configure_app(app);
+    app = endpoints::public_keys::configure_app(app);
+    app = endpoints::store::configure_app(app);
 
     app
 }
